+++ tests/testthat/mod/path/relative/module_d.R
@@ -1,0 +1,5 @@
+box::use(
+  ../to/module_a
+)
+
+module_a$a_fun_a()