+++ tests/testthat/mod/path/relative/module_f.R
@@ -1,0 +1,5 @@
+box::use(
+  ../to/module_a[a_fun_ax]
+)
+
+a_fun_a()