+++ NEWS.md
@@ -1,5 +1,6 @@
 # box.linters (development version)
 
+* [Bug fix] Allow relative box module paths (#110)
 * [Bug fix] Allow multiple `box::use(pkg)` calls (#111)
 
 # box.linters 0.9.1